+++ Changelog.md
@@ -31,6 +31,7 @@
 - Adds VT sequence `DECSNLS` for setting number of lines to display.
 - Adds new CLI command: `contour capture ...` to capture the screen buffer.
 - Adds new CLI command: `contour set profile to NAME` to change the profile on the fly.
+- Adds new CLI command: `contour generate terminfo output OUTPUT_FILE` to create a Contour terminfo file.
 
 ### 0.1.1 (2020-12-31)